コード精査に基づく品質改善: 関数重複解消、テスト修正、Windows互換性、入力検証強化
- wcs_keywords.js を新規作成し isWCSKeyword/makeFITSKeyword の重複を解消
- test_wcs_math.js の Y 座標変換を修正（3件のテスト失敗を解消、55 passed / 0 failed）
- curl パスを /usr/bin/curl → curl に変更（Windows 互換性）
- WCSApplier.js に WCS 値の全フィールド検証を追加
- .gitignore から不要な Python エントリを削除、.build-tmp/ を追加
- 著作権年を 2026 に統一
- リリース ZIP を再ビルド（wcs_keywords.js 追加）
