+++ .gitignore
@@ -1,16 +1,3 @@
-# Python
-__pycache__/
-*.py[cod]
-*$py.class
-*.egg-info/
-dist/
-build/
-*.egg
-
-# 仮想環境
-.venv/
-venv/
-
 # IDE
 .idea/
 .vscode/
@@ -22,13 +9,8 @@ venv/
 .DS_Store
 Thumbs.db
 
-# テスト
-.pytest_cache/
-htmlcov/
-.coverage
-
-# ハンドオーバー
-HANDOVER.md
+# Build
+.build-tmp/
 
 # WCS JSON 出力（ユーザー生成）
 *.wcs.json
