+++ docs/specs.md
@@ -264,7 +264,8 @@ manual-image-solver/
 ├── javascript/
 │   ├── ManualImageSolver.js       # PJSR メイン（ネイティブ Dialog で全操作完結）
 │   ├── WCSApplier.js              # スタンドアロン JSON → WCS 適用
-│   └── wcs_math.js                # WCS 数学関数（PJSR + Node.js 両対応）
+│   ├── wcs_math.js                # WCS 数学関数（PJSR + Node.js 両対応）
+│   └── wcs_keywords.js            # FITS WCS キーワードユーティリティ（PJSR 専用）
 ├── tests/
 │   └── javascript/
 │       ├── test_wcs_math.js       # Node.js 単体テスト（WCS 数学関数）
